Complete A-Packages: Organize into packages
Organise existing classes into suitable packages.
