+++ src/main/java/DukePackage/DukeException.java
@@ -1,3 +1,5 @@
+package DukePackage;
+
 public class DukeException extends Exception {
 
     public DukeException(String msg) {
