+++ src/main/java/TaskPackage/Events.java
@@ -1,3 +1,6 @@
+package TaskPackage;
+
+import DukePackage.DukeException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
